tweet collection versioning and some data readers

+++ main.py
@@ -42,9 +42,10 @@ def main(args):
             # path exists, but it's not a directory
             raise ValueError('ERROR - data path (' + str(args.data_path) + ')already exists, but it is not a directory.')
 
-    # create backup of old file
-    backup_file_name = data_file_name + '.bak'
-    shutil.copy(data_file_name, backup_file_name)
+    # create backup of old file if it exists
+    if os.path.isfile(data_file_name):
+        backup_file_name = data_file_name + '.bak'
+        shutil.copy(data_file_name, backup_file_name)
 
     # get the latest tweets and add them to the data file
     # some tweets may be duplicates for ones recorded earlier, however, some data such as retweet count may be updated
@@ -54,7 +55,7 @@ def main(args):
 
     api = tweepy.API(auth)
     max_tweets = 1000
-    query = 'from:realDonaldTrump'
+    query = args.query
     search_result = [status for status in tweepy.Cursor(api.search, q=query).items(max_tweets)]
 
     with open(data_file_name, 'a') as data_file:
@@ -72,6 +73,7 @@ if __name__ == '__main__':
     parser.add_argument("--config-file", default="config.json",     help="config file name")
     parser.add_argument("--data-path",   default="data",            help="path to data file")
     parser.add_argument("--data-file",   default="trump_dump.json", help="data file name")
+    parser.add_argument("--query",       default="from:realDonaldTrump", help="data file name")
 
     args = parser.parse_args()
